feat: move to action blocking to separate middleware, add comments
