+++ packages/suite/src/middlewares/suite/actionBlockerMiddleware.ts
@@ -1,0 +1,31 @@
+import { MiddlewareAPI } from 'redux';
+import { UI } from 'trezor-connect';
+import { AppState, Action, Dispatch } from '@suite-types';
+
+// actionBlockerMiddleware serves one purpose only, to block certain actions based on activated device's processMode.
+// processMode is set via suiteActions.setProcessMode func
+
+// definition of blocked actions for each process mode
+export const PROCESS_MODE = {
+    'confirm-addr': {
+        blockedActions: [UI.CLOSE_UI_WINDOW], // prevents closing the modal after user confirms the address on a device (Receive tab)
+    },
+} as const;
+
+const actionBlocker = (api: MiddlewareAPI<Dispatch, AppState>) => (next: Dispatch) => async (
+    action: Action,
+): Promise<Action> => {
+    const prevState = api.getState();
+
+    // block actions restricted by device's process mode
+    const processMode = prevState.suite.device?.processMode;
+    if (processMode && action.type in PROCESS_MODE[processMode]?.blockedActions) {
+        return action;
+    }
+
+    // pass action to reducers
+    next(action);
+    return action;
+};
+
+export default actionBlocker;
