+++ packages/suite/src/middlewares/suite/index.ts
@@ -5,5 +5,6 @@ import analytics from './analyticsMiddleware';
 import buttonRequest from './buttonRequestMiddleware';
 import events from './eventsMiddleware';
 import metadata from './metadataMiddleware';
+import actionBlocker from './actionBlockerMiddleware';
 
-export default [log, redirect, suite, analytics, buttonRequest, events, metadata];
+export default [log, redirect, suite, analytics, buttonRequest, events, metadata, actionBlocker];
